requirements.txt
librosa
transformers
torch
pyannote.audio
google-generativeai
langchain
langchain-google-genai
langchain-community
sentence-transformers
faiss-cpu
scikit-learn
python-dotenv